backend/requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
supabase==2.3.4

# Image Processing
pillow==10.2.0
imagehash==4.3.1
opencv-python==4.9.0.80
numpy==1.26.3
scipy==1.12.0
scikit-image==0.22.0

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
aiofiles==23.2.1

# Background Tasks
celery==5.3.6
redis==5.0.1

# CORS
starlette==0.35.1
